TURMERIC-1582: Fix failing tests to handle change in data aggregation.
The runtime sends an accumlated value for both Count and Response time.
The four failing tests were sending the data non accumlated.  So when
the resolve method in the provider was called, and the difference was
done, it was making the second or third return zero for both the count
and the response time.   Solution to this was to make sure the tests
accumlated the count and response time when they created a new AverageValue
object, and passed in the values.  This then allows the difference
method to correctly reduce the value, and correctly calculate the
value to store.