TURMERIC-1582: Refactor cassandra storage provider to leverage SnapshotLogger
The runtime has an abstract class called SnapshotLogger, this provides
a general way for metric collectors to be implemented.  It provides pre
and post hooks that can be used by a provider.  It also provides a standard
way to handle difference.   The an AbstractDiffereceProvider is also
created, this provides the ability to get the difference from a previous
snapshot.  It also implements the DifferenceStorageProvider interface to
allow reseting of service metrics instead of accumulating them all the
time.

The Cassandra Storage provider has been updated to use this new
implementation, and extrandeous or dead code has been removed.  All existing
unit tests pass as well with this implementation.